Update patent module API and middleware
