+++ FusionIIIT/Fusion/urls.py
@@ -65,7 +65,7 @@ urlpatterns = [
     url(r'^recruitment/', include('applications.recruitment.urls')),
     url(r'^examination/', include('applications.examination.urls')),
     url(r'^otheracademic/', include('applications.otheracademic.urls')),
-    url(r'^patentsystem/', include('applications.patent_system.urls')),
+    url(r'^patentsystem/', include('applications.patent_system.api.urls')),
 
     path(
         'password-reset/',
